Fix: [Server][Client] iPhone Safari での再生フリーズを修正（セッションタイムアウト拡大と自動セッション復旧）
- サーバ: SESSION_TIMEOUT を 10 秒から 120 秒へ拡大し、モバイル Safari で前面再生中でも
  発生する keep-alive の一時的な断絶（Wi-Fi 省電力・タイマー抑制）でセッションが
  破棄されないようにする
- クライアント: セッション消滅からの自動復旧を追加
  - hls.js の fatal エラーと keep-alive の HTTP 422 を検知し、新しい cache_key で
    プレイリストを再取得して同一 session_id のセッションをサーバ側で再作成する
  - 復旧中の旧セッション残骸エラーは cache_key で判別して試行回数を消費しない
  - 30 秒間に 3 回失敗した場合は自動復旧を断念してユーザーへ通知する

+++ server/app/streams/VideoStream.py
@@ -82,7 +82,9 @@ class VideoStream:
 
     # 録画視聴セッションが再生されていない場合にタイムアウトするまでの時間 (秒)
     # この時間が経過すると、録画視聴セッションのインスタンスは自動的に破棄される
-    SESSION_TIMEOUT: ClassVar[float] = float(10)  # 10 秒
+    # モバイル Safari では前面再生中でも、Wi-Fi 省電力やタイマー抑制により keep-alive が 10 秒以上途切れることがある
+    # ため、余裕を持たせている。トレードオフとして、視聴終了後もエンコーダが最大 120 秒残る
+    SESSION_TIMEOUT: ClassVar[float] = float(120)  # 120 秒
 
     # 一度でも読み取られた HLS セグメントの最大保持数
     MAX_READED_SEGMENTS: ClassVar[int] = 10
